Projections of a bipartite network

Projections of a bipartite network

projections(px = NULL, py = NULL, f = "from", t = "to",
  v = "value", ax = 4, ay = 4, tol = 0.01, tbl = TRUE,
  compute = "both")

Arguments

px
tibble/data.frame or sparse/dense matrix containing proximity measures for the elements of set X (e.g. proximity_x from proximity())

py
tibble/data.frame or sparse/dense matrix containing proximity measures for the elements of set Y (e.g. proximity_y from proximity())

f
column with the "from" side in a relation between elements of the same set (applies only if d is a tibble/data.frame)

t
column with the "to" side in a relation between elements of the same set (applies only if d is a tibble/data.frame)

v
column with a metric of the relation between the elements of the same set (applies only if d is a tibble/data.frame)

ax
average number of connections for the projection of X (default set to 4)

ay
average number of connections for the projection of Y (default set to 4)

tol
tolerance for proximity variation on each iteration until obtaining the desired average number of connections (default set to 0.01)

tbl
when set to TRUE the output will be a tibble instead of a graph (set to TRUE by default)

compute
set to "both" by default, it can also be "x" or "y" to obtain one projection of the bipartite network

References

For more information see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

and the references therein.

Examples

projections(
  px = binet_output$proximity$proximity_x,
  py = binet_output$proximity$proximity_y,
  f = "from",
  t = "to",
  v = "value",
  tol = 0.25,
  tbl = TRUE
)

#> Computing projection of X...

#> 0 threshold...

#> 0.25 threshold...

#> 0.5 threshold...

#> 0.75 threshold...

#> 1 threshold...

#> no threshold achieves the avg number of connections for X projection, returning maximum spanning tree

#> Computing Y projection...

#> 0 threshold...

#> 0.25 threshold...

#> 0.5 threshold...

#> 0.75 threshold...

#> 1 threshold...

#> no threshold achieves the avg number of connections for Y projection, returning maximum spanning tree

#> $network_x
#> # A tibble: 157 x 3
#>    from  to     value
#>    <chr> <chr>  <dbl>
#>  1 irn   afg   -0.5  
#>  2 npl   afg   -0.214
#>  3 syr   afg   -0.357
#>  4 ago   ben   -0.339
#>  5 ago   per   -0.358
#>  6 ago   sen   -0.308
#>  7 irn   alb   -0.375
#>  8 sur   and   -0.130
#>  9 bhr   ant   -0.312
#> 10 sur   ant   -0.304
#> # … with 147 more rows
#> 
#> $network_y
#> # A tibble: 990 x 3
#>    from  to     value
#>    <chr> <chr>  <dbl>
#>  1 9614  0101  -0.579
#>  2 0102  0104  -0.423
#>  3 0102  0201  -0.5  
#>  4 0103  0505  -0.647
#>  5 1102  0105  -0.417
#>  6 4803  0105  -0.417
#>  7 0106  1302  -0.514
#>  8 0201  0208  -0.619
#>  9 0201  0210  -0.524
#> 10 0201  0409  -0.571
#> # … with 980 more rows
#>