Proximity

TBDs

proximity(
  balassa_index,
  balassa_sum_source,
  balassa_sum_target,
  source = "source",
  target = "target",
  value = "value",
  compute = "both"
)

Arguments

balassa_index
a data frame (e.g. the output from balassa_index()).

balassa_sum_source
numeric vector or tibble/data.frame containing the Balassa sum for elements of set X (e.g. balassa_sum_source from complexity_measures()).

balassa_sum_target
numeric vector or tibble/data.frame containing the Balassa sum for elements of set Y (e.g. balassa_sum_target from complexity_measures()).

source
a column with the elements of set X (applies only if data is a data frame).

target
a column with the elements of set Y (applies only if data is a data frame).

value
a column with some metric of the relation between the elements of X and Y (applies only if data is a data frame).

compute
which proximity to compute. By default is "both" (both proximities) but it can also be "source" or "target".

Details

TBD

References

For more information see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

and the references therein.

Examples

proximity(
  balassa_index = binet_output$balassa_index,
  balassa_sum_source = binet_output$complexity_measures$balassa_sum_source,
  balassa_sum_target = binet_output$complexity_measures$balassa_sum_target
)

#> $proximity_source
#> # A tibble: 10,432 x 3
#>    source target   value
#>    <chr>  <chr>    <dbl>
#>  1 ago    afg    0.0962 
#>  2 alb    afg    0.219  
#>  3 arg    afg    0.120  
#>  4 aus    afg    0.111  
#>  5 aut    afg    0.00690
#>  6 bdi    afg    0.0714 
#>  7 bel    afg    0.00938
#>  8 ben    afg    0.107  
#>  9 bfa    afg    0.0816 
#> 10 bgr    afg    0.0945 
#> # … with 10,422 more rows
#> 
#> $proximity_target
#> # A tibble: 417,822 x 3
#>    source target value
#>    <chr>  <chr>  <dbl>
#>  1 0102   0101   0.346
#>  2 0103   0101   0.263
#>  3 0104   0101   0.421
#>  4 0106   0101   0.2  
#>  5 0201   0101   0.381
#>  6 0203   0101   0.474
#>  7 0204   0101   0.158
#>  8 0205   0101   0.368
#>  9 0206   0101   0.316
#> 10 0207   0101   0.368
#> # … with 417,812 more rows
#>