Projections of a bipartite network

TBD

projections(
  proximity_source,
  proximity_target,
  source = "source",
  target = "target",
  value = "value",
  avg_links = 4,
  tolerance = 0.05,
  compute = "both"
)

Arguments

proximity_source
a data frame containing proximity values for the elements of set X (e.g. proximity_source from proximity())

proximity_target
a data frame containing proximity values for the elements of set Y (e.g. proximity_target from proximity())

source
a column with the elements of set X (applies only if data is a data frame).

target
a column with the elements of set Y (applies only if data is a data frame).

value
a column with the proximity values between the elements of X or Y (applies only if data is a data frame).

avg_links
average number of connections for the projection of X (default set to 4)

tolerance
tolerance for proximity variation on each iteration until obtaining the desired average number of connections (default set to 0.05)

compute
which projection to compute. By default is "both" (both projections) but it can also be "source" or "target".

Details

TBD

References

For more information see:

Hausmann R, Hidalgo C, Bustos S, Coscia M, Simoes A, Yildirim M (2014). The atlas of economic complexity: Mapping paths to prosperity. MIT Press. doi: 10.7551/mitpress/9647.001.0001 .

and the references therein.

Examples

projections(
  proximity_source = binet_output$proximity$proximity_source,
  proximity_target = binet_output$proximity$proximity_target
)

#> computing target projection...

#> --------------------------------------------------

#> 0 threshold...

#> 0.05 threshold...

#> 0.1 threshold...

#> 0.15 threshold...

#> 0.2 threshold...

#> 0.25 threshold...

#> 0.3 threshold...

#> 0.35 threshold achieves the avg number of connections

#> computing target projection...

#> --------------------------------------------------

#> 0 threshold...

#> 0.05 threshold...

#> 0.1 threshold...

#> 0.15 threshold...

#> 0.2 threshold...

#> 0.25 threshold...

#> 0.3 threshold...

#> 0.35 threshold...

#> 0.4 threshold...

#> 0.45 threshold...

#> 0.5 threshold...

#> 0.55 threshold...

#> 0.6 threshold...

#> 0.65 threshold...

#> 0.7 threshold achieves the avg number of connections

#> $network_source
#> # A tibble: 272 x 3
#>    source target value
#>    <chr>  <chr>  <dbl>
#>  1 irn    syr    0.344
#>  2 irn    alb    0.375
#>  3 irn    tur    0.348
#>  4 irn    afg    0.5  
#>  5 npl    afg    0.214
#>  6 syr    afg    0.357
#>  7 ago    ben    0.339
#>  8 ago    khm    0.308
#>  9 ago    moz    0.304
#> 10 ago    per    0.358
#> # … with 262 more rows
#> 
#> $network_target
#> # A tibble: 1,800 x 3
#>    source target value
#>    <chr>  <chr>  <dbl>
#>  1 9614   4202   0.529
#>  2 9614   0101   0.579
#>  3 0102   0201   0.5  
#>  4 0102   0104   0.423
#>  5 0103   0505   0.647
#>  6 1102   1005   0.5  
#>  7 1102   0105   0.417
#>  8 4803   4707   0.615
#>  9 4803   0105   0.417
#> 10 0106   1302   0.514
#> # … with 1,790 more rows
#>